+++ src/main/scala/edg_ide/psi_edits/LiveTemplateConnect.scala
@@ -4,10 +4,12 @@ import com.intellij.codeInsight.template.impl.TemplateState
 import com.intellij.openapi.command.WriteCommandAction.writeCommandAction
 import com.intellij.openapi.project.Project
 import com.intellij.psi.PsiElement
-import com.jetbrains.python.psi.PyClass
+import com.intellij.psi.util.PsiTreeUtil
+import com.jetbrains.python.psi.{PyClass, PyFunction, PyStatement}
 import edgir.elem.elem
 import edg.util.Errorable
-import edg_ide.util.{ConnectBuilder, PortConnects, exceptable}
+import edg_ide.util.ExceptionNotifyImplicits.{ExceptNotify, ExceptOption}
+import edg_ide.util.{ConnectBuilder, PortConnects, exceptable, requireExcept}
 
 object LiveTemplateConnect {
   // Creates an action to start a live template to insert the connection
@@ -22,8 +24,18 @@ object LiveTemplateConnect {
   ): Errorable[() => Unit] = exceptable {
 
     val movableLiveTemplate = new MovableLiveTemplate(actionName) {
+      // TODO startTemplate should be able to fail - Errorable
       override def startTemplate(caretEltOpt: Option[PsiElement]): InsertionLiveTemplate = {
-        ???
+        val caretElt = caretEltOpt.exceptNone("no elt at caret")
+        val caretStatement = InsertAction.snapInsertionEltOfType[PyStatement](caretElt).get
+        val containingPsiFn = PsiTreeUtil
+          .getParentOfType(caretStatement, classOf[PyFunction])
+          .exceptNull(s"caret not in a function")
+        val containingPsiClass = PsiTreeUtil
+          .getParentOfType(containingPsiFn, classOf[PyClass])
+          .exceptNull(s"caret not in a class")
+        requireExcept(containingPsiClass == contextClass, s"caret not in class of type ${contextClass.getName}")
+        caretStatement
       }
     }
