+++ src/main/scala/edg_ide/psi_edits/TemplateUtils.scala
@@ -3,6 +3,10 @@ package edg_ide.psi_edits
 import com.intellij.codeInsight.template.{Template, TemplateEditingAdapter}
 import com.intellij.codeInsight.template.impl.TemplateState
 import com.intellij.psi.PsiElement
+import com.intellij.psi.util.PsiTreeUtil
+import com.jetbrains.python.psi.{PyClass, PyFunction, PyStatement}
+import edg_ide.util.ExceptionNotifyImplicits.ExceptNotify
+import edg_ide.util.requireExcept
 
 import scala.collection.mutable
 
@@ -34,6 +38,19 @@ object TemplateUtils {
       }
     }
     templateState.update() // update to end the template
+  }
+
+  // given some caret position, returns the top-level statement if it's valid for statement insertion
+  def getInsertionStmt(caretElt: PsiElement, requiredContextClass: PyClass): Option[PyStatement] = {
+    val caretStatement = InsertAction.snapInsertionEltOfType[PyStatement](caretElt).get
+    val containingPsiFn = PsiTreeUtil
+      .getParentOfType(caretStatement, classOf[PyFunction])
+    if (containingPsiFn == null) return None
+    val containingPsiClass = PsiTreeUtil
+      .getParentOfType(containingPsiFn, classOf[PyClass])
+    if (containingPsiClass == null) return None
+    if (containingPsiClass != requiredContextClass) return None
+    Some(caretStatement)
   }
 }
